% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{base_cep}
\alias{base_cep}
\title{Georeferência para CEP das unidades prisionais}
\format{
:
\describe{
\item{cep}{CEP da unidade prisional}
\item{lat}{latitude}
\item{long}{longitude}
}
}
\source{
\url{https://www.cepaberto.com/api_key}
}
\usage{
base_cep
}
\description{
A base foi baixada usando a API do CEP Aberto, mais especificamente
a função \code{\link[cepR:busca_cep]{cepR::busca_cep()}}.
}
\keyword{datasets}
