% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\encoding{UTF-8}
\name{base_perfilpessoal}
\alias{base_perfilpessoal}
\title{Variáveis relativas ao perfil pessoal da população prisioal}
\format{
:
\describe{
\item{id_origem_sisdepen}{ID de qual edição do SISDEPEN a base
corresponde}
\item{id_ano_sisdepen}{Ano em que as informações foram enviadas}
\item{id_mes_sisdepen}{Qual o período se refere, em geral sendo junho,
dezembro ou "único", nos anos em que era enviado apenas uma vez}
\item{nome_estabelecimento, sigla_uf e cod_ibge}{Nome do estabelecimento prisional,
UF de onde está o estabelecimento e o código IBGE (7 dígitos) do município onde
está localizado}
\item{possui_registro_raca, possui_registro_idade, possui_registro_pcd, possui_registro_estadocivi,
possui_registro_procedencia}{Informa se a unidade possui condições de obter em seus registros
as informações respectivas (raca, idade, etc) e em que grau (de todas as pessoas ou de parte delas).
Pode ser relevante para saber o grau de confiança que se pode ter dessa informação.}
\item{quantidade_raca_)}{Informa a quantidade de pessoas por cada uma das raças ou etnias
de cada um dos gêneros (_mas ou _fem).
Não está claro se o dado é obtido através apenas de autodeclaração ou se essa
categorização é feita com base em outra maneira (o que consta de cadastro público ou
interpretação da própria unidade, por exemplo).

Pode ser:
\itemize{
\item \emph{\emph{branca}}
\item \emph{\emph{preta}}
\item \emph{\emph{parda}}
\item \emph{\emph{indigena}}
\item \emph{\emph{naoinformada}}
}

}
\item{quantidade_idade_)}{Divisão da população prisional existente na unidade de
acordo com faixas etárias, para cada um dos gêneros masculino e feminino. As faixas
etárias existentes são:

-\emph{18 a 24} anos
-\emph{25 a 29} anos
-\emph{30 a 34} anos
-\emph{35 a 45} anos
-\emph{46 a 60} anos
-\emph{61 a 70} anos
-\emph{mais de 70} anos
-\emph{naoinformada}

}
\item{quantidade_procedencia_)}{Divisão da população de acordo com a zona
de procedência da pessoa, dividido entre masculino e feminino. Pode ser:

-\emph{interior}
-\emph{rmetropolitana} região metropolitana
-\emph{rural} zona rural

}
}
}
\source{
https://www.gov.br/depen/pt-br/sisdepen/mais-informacoes/bases-de-dados
}
\usage{
base_perfilpessoal
}
\description{
Seleção das principais variáveis relativas à população prisional contidas na
primeira parte do item 5 do formulário do SISDEPEN. Em geral, dizem respeito
às características pessoais tais como idade, raça, gênero, nacionalidade,
existência ou não de deficiência física, etc.
}
\keyword{datasets}
