% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\encoding{UTF-8}
\name{base_populacao}
\alias{base_populacao}
\title{Variáveis relativas à população prisional}
\format{
:
\describe{
\item{id_origem_sisdepen}{ID de qual edição do SISDEPEN a base
corresponde}
\item{id_ano_sisdepen}{Ano em que as informações foram enviadas}
\item{id_mes_sisdepen}{Qual o período se refere, em geral sendo junho,
dezembro ou "único", nos anos em que era enviado apenas uma vez}
\item{nome_estabelecimento, sigla_uf e cod_ibge}{Nome do estabelecimento prisional,
UF de onde está o estabelecimento e o código IBGE (7 dígitos) do município onde
está localizado}
\item{quantidade_(situação)\emph{(regime)}(origem)_(gênero)}{Variáveis indicando a quantidade
de pessoas custodiadas no momento do envio das respostas ao SISDEPEN. Não se confunde
com a quantidade de "vagas", já que pode ser menor ou - infelizmente, na maior parte
dos casos - muito maior do que o esperado.

Para fins do SISDEPEN a pessoa deixa de ser considerada provisória quando há sentença,
independente de ainda cabe recurso (ou seja, mesmo que não haja o trânsito em julgado).
Caso a pessoa esteja condenada mas também possua prisão provisória por outro processo ela
será considerada, para fins do SISDEPEN, condenada.

Essas variáveis estão em uma estrutura divididos pela situação processual, regime,
origem da prisão (qual Justiça) e gênero:

\emph{Situação processual}
Pode ser:
\enumerate{
\item \emph{presoprovisorio} Indicando serem provisórios (sem sentença condenatória)
\item \emph{sentenca} Indicando serem já sentenciados
\item \emph{medseg} Medida de segurança
}

\emph{Regime de pena}
\enumerate{
\item Se \emph{sentenca} pode variar entre \emph{ra} (regime aberto), \emph{rsa} (regime semiaberto) ou
\emph{rf} (regime fechado)
\item Se \emph{medseg} pode ser \emph{ambulatorial} (tratamento ambulatorial) ou \emph{internacao}
}

\emph{Justiça de origem}
\enumerate{
\item \emph{jusestadual} : justiça estadual
\item \emph{jusfederal} : justiça federal
\item \emph{outrajus} : outras justiças como Trabalhista, Eleitoral ou cível.
}

}
}
}
\source{
https://www.gov.br/depen/pt-br/sisdepen/mais-informacoes/bases-de-dados
}
\usage{
base_populacao
}
\description{
Seleção das principais variáveis relativas à população prisional contidas no
item 4 do formulário do SISDEPEN. Em geral, dizem respeito às caractarísticas
população custodiada na unidade.
}
\keyword{datasets}
