% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/criar_blocos.R
\encoding{UTF-8}
\name{blocos}
\alias{blocos}
\alias{carrega_e_aplica}
\alias{arruma_bloco_estabelecimento}
\alias{arruma_bloco_populacao}
\alias{arruma_bloco_perfilpessoal}
\alias{arruma_bloco_perfilcriminal}
\title{Criar bases a partir dos blocos de perguntas do SISDEPEN}
\usage{
carrega_e_aplica(funcao, lista)

arruma_bloco_estabelecimento(base)

arruma_bloco_populacao(base)

arruma_bloco_perfilpessoal(base)

arruma_bloco_perfilcriminal(base)
}
\arguments{
\item{funcao}{Uma das funções arruma_bloco_ a ser indicada}

\item{lista}{Lista de arquivos a serem abertos. Se não for informada,
irá buscar tudo que estiver na pasta /data-raw/sisdepen}

\item{base}{Nas funções arruma_bloco_ recebe a base do DEPEN a ser
"fatiada"}
}
\value{
Tibble
}
\description{
As milhares de colunas do SISDEPEN foram divididas nesse pacote em
"bloco", de acordo com a pertinência temática de cada uma delas.

Para cada bloco foi gerada uma função que seleciona as colunas
pertinentes a ela.

A função carrega_e_aplica() recebe uma das funções e a lista de
arquivos .xlsx já baixados para carregar e aplicar a cada um deles,
formando ao final uma base temática.
}
