% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/carrega_base.R
\encoding{UTF-8}
\name{carregar_base_sisdepen}
\alias{carregar_base_sisdepen}
\title{Carrega base do SISDEPEN}
\usage{
carregar_base_sisdepen(path, padrao = "guess")
}
\arguments{
\item{path}{Caminho para a base. Deve conter o ano e mês de origem.
Deve ser \code{.xlsx}}

\item{padrao}{"guess" por default. Indica o tipo de coluna padrão. Se
tiver problema de compatibilidade pode ser uma boa colocar "text".}
}
\value{
Tibble
}
\description{
Carrega a base indicada de modo que todas as colunas sejam strings e
que a coluna de contendo o nome do estabelecimento seja renomeada para
\code{nome_estabelcimento} e UF seja renomeada para \code{sigla_uf}.

Além disso, será gerado um identificador para a base na coluna
\code{id_origem_sisdepen} de acordo com o ano e período a que ela se referir.
}
