% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auxiliares.R
\name{checa}
\alias{checa}
\title{Wrap de stringr}
\usage{
checa(string, pattern, ignore_case = TRUE, ...)
}
\arguments{
\item{string}{Variável a ser analisada}

\item{pattern}{Regex}

\item{ignore_case}{TRUE por padrão.}

\item{...}{Outros valores para serem passados no stringr::regex}
}
\value{
Logical
}
\description{
Atalho para stringr::str_detect(), considerando o ignore.case = TRUE
}
