% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auxiliares.R
\encoding{UTF-8}
\name{conta_valores}
\alias{conta_valores}
\title{Conta os valores das colunas}
\usage{
conta_valores(base, coluna = "*")
}
\arguments{
\item{base}{A base a ser utilizada}

\item{coluna}{A expressão a ser usada para selecionar a coluna.}
}
\value{
Lista
}
\description{
Conta os valores das colunas
}
