% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auxiliares.R
\encoding{UTF-8}
\name{ler_colunas}
\alias{ler_colunas}
\title{Ler os nomes das colunas e colocar na área de transferência}
\usage{
ler_colunas(base)
}
\arguments{
\item{base}{}
}
\value{
Vetor de character contendo as colunas da base.
}
\description{
Em caso de necessitar manipular bases com muitas colunas,
a função irá carregar e imprimir em formato amigável e colocar na
área de transferência para que seja colada com CNTR+V.

Útil para colocar as colunas no regex101.cim
}
