% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auxiliares.R
\encoding{UTF-8}
\name{numero_para_sim_nao}
\alias{numero_para_sim_nao}
\title{Transformar números em sim/não}
\usage{
numero_para_sim_nao(valor, transforma_na = FALSE, logico = FALSE)
}
\arguments{
\item{valor}{Vetor a ser transformado}

\item{transforma_na}{Se TRUE vai transformar NA em Não}

\item{logico}{Se TRUE vai transformar Sim em TRUE e Não em NA}
}
\value{
Vetor
}
\description{
Transformar números em sim/não
}
