% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gera_bloco1.R
\encoding{UTF-8}
\name{renomeia_bloco_estabelecimento}
\alias{renomeia_bloco_estabelecimento}
\title{Gera bloco de colunas sobre o estabelecimento}
\usage{
renomeia_bloco_estabelecimento(base)
}
\arguments{
\item{base}{}
}
\value{
Tibble
}
\description{
Gera bloco de colunas sobre o estabelecimento
}
