% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auxiliares.R
\encoding{UTF-8}
\name{renomeia_coluna}
\alias{renomeia_coluna}
\title{Renomeia colunas}
\usage{
renomeia_coluna(base, nome, coluna, fixa = "cod_ibge")
}
\arguments{
\item{base}{Tibble da base}

\item{nome}{Nome que será recebido}

\item{coluna}{Expressão que identificará a coluna}

\item{fixa}{Por padrão, "sigla_uf". Identifica após qual coluna a nova será
recolocada.}
}
\value{
Tibble renomeada
}
\description{
Aplica a função dplyr::rename() à uma base, mudando o nome da coluna indicada.
Será aplicada a função dplyr::matches() para identificar a coluna.

Ao final, a coluna nova será
}
