% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{base_perfilcriminal}
\alias{base_perfilcriminal}
\title{Base estabelecimento....}
\format{
:
\describe{
\item{cep}{CEP da unidade prisional}
\item{lat}{latitude}
\item{long}{longitude}
}
}
\source{
SISDEPEN
}
\usage{
base_perfilcriminal
}
\description{
A base foi baixada usando a API do CEP Aberto, mais especificamente
a função \code{\link[cepR:busca_cep]{cepR::busca_cep()}}.
}
\keyword{datasets}
