% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\encoding{UTF-8}
\name{base_perfilcriminal}
\alias{base_perfilcriminal}
\title{Variáveis relativas ao perfil criminal}
\format{
:
\describe{
\item{id_origem_sisdepen}{ID de qual edição do SISDEPEN a base
corresponde}
\item{id_ano_sisdepen}{Ano em que as informações foram enviadas}
\item{id_mes_sisdepen}{Qual o período se refere, em geral sendo junho,
dezembro ou "único", nos anos em que era enviado apenas uma vez}
\item{nome_estabelecimento, sigla_uf e cod_ibge}{Nome do estabelecimento prisional,
UF de onde está o estabelecimento e o código IBGE (7 dígitos) do município onde
está localizado}
\item{existe_registro_incidenciapenal}{Informa se o estabelecimento tem alguma
forma de controlar o registro sobre a incidência penal}
\item{metodologia_registro_incidenciapenal}{Qual a metodologia utilizada para o
registro da incidência penal. Permite saber o grau de atualização dos dados e,
com isso, mensurar o quanto é possível confiar na informação.}
\item{quantidade_com_informacao_incidencia e quantidade_sem_informacao_incidencia}{Quantas
pessoas estão ou não com essa informação. Junto à variável de metodologia permite aferir
o quão atualizados estão os dados}
\item{quantidade_incidencia_(crime)(modalidade)}{Variáveis indicando a quantidade
de incidências penais para cada um dos crimes. Como dito acima, uma pessoa pode possuir
mais de uma incidência penal.

Foram divididos de acordo com a categoria do crime, podendo ser divididos em modalidade
quando for o caso.

\emph{Crimes contra a pessoa}
\itemize{
\item \emph{homicídio}: podendo ser, _culposo, _simples ou _qualificado
\item \emph{aborto}
\item \emph{lesaocorporal}
\item \emph{violenciadomestica}
\item \emph{sequestrocarcereprivado} previsto no art. 148 do Código Penal
\item \emph{outros_contrapessoa} outros crimes contra a pessoa não listados acima
}

\emph{Crimes contra o patrimônio}
\itemize{
\item \emph{furto} : podendo ser _simples ou _qualificado
\item \emph{roubo} : podendo ser _simples ou _qualificado
\item \emph{latrocinio} : também conhecido como roubo seguido de morte
\item \emph{extorsao} : art. 158 do CP
\item \emph{extorsaomedsequestro} : extorsão mediante sequestro (art. 159 do CP)
\item \emph{apropindebita} : apropriação indébita (art. 168)
\item \emph{apropindebita_previdenciaria} : apropriação indébita previdenciária (art. 168-A)
\item \emph{estelionato}
\item \emph{receptacao} receptação, podendo ser também _qualificada
\item \emph{outros_contrapatrimonio}
}

\emph{Contra a dignidade sexual}
\itemize{
\item \emph{estupro}
\item \emph{estupro_vulneravel}
\item \emph{atentadoviolentopudor}
\item \emph{corrupcaomenores}
\item \emph{traficopessoasexploracaosexua} podendo ser _interno ou _internacional
\item \emph{outros_dignidadesexual}
}

\emph{Contra a fé pública e contra administração pública}
\itemize{
\item \emph{quadrilha} formação de quadrilha ou bando (art. 288)
\item \emph{moedafalsa}
\item \emph{falsificacaodocpublicos} falsificação de papéis, seos, sinal e documentos
públicos (art. 293 a 297)
\item \emph{falsidadeideologica}
\item \emph{usodocfalso} uso de documento falso
\item \emph{peculato}
\item \emph{concussão} concussão e excesso de exação (art. 316)
\item \emph{corrupcao_ativa}
\item \emph{corrupcao_passiva}
\item \emph{contrabando} contrabando ou descaminho
}

\emph{Drogas}
\itemize{
\item \emph{drogas_trafico} Art. 33 da Lei de Drogas atual (e art. 12 da anterior)
\item \emph{drogas_trafico_internacional} Art. 33 com 40, I da Lei de Drogas atual
\item \emph{drogas_associacao} Associação para o tráfico, art. 35 da Lei de Drogas
}

\emph{Armas} (crimes do Estatuto do Desarmamento)
\itemize{
\item \emph{armas_usopermitido} Porte ilegal de arma de fogo de uso permitido (art. 14)
\item \emph{armas_usorestrito} Posse ou porte ilegal de arma de uso restrito (art. 16)
\item \emph{armas_disparo} Disparo de arma de fogo (art. 15)
\item \emph{armas_comercioilegal} Comércio ilegal de armas de fogo (art. 17)
\item \emph{armas_traficointernacional} Tráfico internacional de arma de fogo (art. 18)
}

\emph{Outros crimes de legislação específica}
\itemize{
\item \emph{transito_homicidioculposo} Homicídio culposo na condução de veículo automotor (art. 302 do Código de Trânsito)
\item \emph{transito_outros} Outros crimes previstos no Código de Trânsito Braileiro
\item \emph{estatutocriancaadolescente} Crimes previstos no ECA (Estatuto da Criança e Adolescente).
Note que em regra são crimes cometidos \emph{contra} crianças e adolescentes e não \emph{por} crianças e adolecentes.
Nesse último caso essas pessoas não estão no sistema penitenciário pois se aplica legislação específica.
\item \emph{contrameioambiente} Crimes previstos na Lei dos Crimes Ambientais (Lei 9.605)
\item \emph{genocidio} Crime previsto na Lei 2.889/56
\item \emph{tortura} Crime previsto na Lei 9.455/97
}

}
}
}
\source{
https://www.gov.br/depen/pt-br/sisdepen/mais-informacoes/bases-de-dados
}
\usage{
base_perfilcriminal
}
\description{
Seleção das principais variáveis relativas à população prisional contidas na
segunda parte do item 5 do formulário do SISDEPEN. Em geral, dizem respeito
ao crime pelo qual a pessoa está presa (seja em caráter provisório ou não).

A incidência penal é considerada aquela classificada no auto de prisão em flagrante,
instauração de inquérito ou denúncia caso seja um preso provisório. No caso de
sentenciado, considera-se a capitulação dada na sentença ou acórdão.

\emph{\emph{ATENÇÃO}}: O SISDEPEN determina que as incidências serão registradas de \emph{forma cumulativa}: se uma pessoa foi condenada
por homicídio simples e por roubo simples, deve ser lançado um registro em
homicídio simples e outro registro em roubo simples, da mesma forma, se a pessoa
foi condenada por roubo simples e aguarda julgamento por tráfico de drogas, deve
ser lançado registro nos dois tipos penais.
}
\keyword{datasets}
