% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/criar_blocos.R
\encoding{UTF-8}
\name{blocos}
\alias{blocos}
\alias{carrega_e_aplica}
\alias{arruma_bloco_estabelecimento}
\alias{arruma_bloco_populacao}
\alias{arruma_bloco_perfilpessoal}
\alias{arruma_bloco_perfilcriminal}
\title{Criar bases a partir dos blocos de perguntas do SISDEPEN}
\usage{
carrega_e_aplica(funcao, lista)

arruma_bloco_estabelecimento(base)

arruma_bloco_populacao(base)

arruma_bloco_perfilpessoal(base)

arruma_bloco_perfilcriminal(base)
}
\arguments{
\item{base}{}
}
\value{

}
\description{
Criar bases a partir dos blocos de perguntas do SISDEPEN
}
